+++ AnugerahBackend/StokBarang/Model/BPStokDetilModel.cs
@@ -15,10 +15,10 @@ namespace AnugerahBackend.StokBarang.Model
         public string Tgl { get; set; }
         public string Jam { get; set; }
 
-        public long QtyIn { get; set; }
+        public decimal QtyIn { get; set; }
         public decimal NilaiHpp { get; set; }
 
-        public long QtyOut { get; set; }
+        public decimal QtyOut { get; set; }
         public decimal HargaJual { get; set; }
     }
 }
